fix(launch.json): avoid exception when .py changes
Before this commit, while running a debugging session, if a python file
changes then the odoo instance emits an exception.
This is due to the fact that when relaunched the __package__ global is
no more defined.

After this commit:
- the debugging configurations do not use the python module "odoo" but
  instead they launch directly the python file "odoo-bin".

This may not be the best fix, but for now it is working.